Allow MC version to be changed using the dropdown.
I have made an executive decision to remove the API and Forge selectors
from the download page. They serve little to no purpose because people
coming to the downloads page will be looking for downloads for Minecraft
versions and nothing else. Having to select a specific MC version, API
version AND Forge version to get builds for SF to show up is not exactly
user friendly.

We can display the tags next to the downloads if needs be, but that's a
future job, probably for a full front end replacement.

This also throws up a potential SOAD issue (or supplied tag issue) --
the MC tag for 1.17 pulls up versions 1.17, 1.17-RC1, 1.17.1, 1.17.1-PRE1,
1.17.1-RC1, 1.17.1-RC2 -- but that needs looking at from the SOAD side.

Also moved some SCSS out of the Vue file.